+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-* List pending changes here.
+* Add Docker Compose transport.
 
 ### 0.1.0 - 2018/Oct/23